Hoist `brand` to `advanced-types`
